refactor storageManager
Refactor StorageManager to work with repository e.g. GitLab
