+++ sandbox_scripts/QualiEnvironmentUtils/SnapshotManager.py
@@ -3,9 +3,9 @@ from sandbox_scripts.QualiEnvironmentUtils.StorageManager import StorageManager
 from sandbox_scripts.QualiEnvironmentUtils.QualiUtils import QualiError
 '''
 class SnapshotManager(object):
-    def __init__(self, storage_client):
-        self.storage_client = storage_client
-        self.config_files_root = self.storage_client.get_configs_root()
+    def __init__(self, storage_mgr):
+        self.storage_mgr = storage_mgr
+        self.config_files_root = self.storage_mgr.get_configs_root()
     # ----------------------------------
     # Is this Sandbox originates from a snapshot Blueprint?
     # ----------------------------------
@@ -17,7 +17,7 @@ class SnapshotManager(object):
             env_dir = self.config_files_root + '/Snapshots/' + self.sandbox.Blueprint_name
 
         env_dir = env_dir.replace(' ', '_')
-        return self.storage_client.dir_exist(env_dir)
+        return self.storage_mgr.dir_exist(env_dir)
 
     # ----------------------------------
     # delete file name on storage
@@ -26,11 +26,11 @@ class SnapshotManager(object):
 
         env_dir = self.config_files_root + '/Snapshots/' + fileName
         env_dir = env_dir.replace(' ', '_')
-        self.storage_client.delete(env_dir)
+        self.storage_mgr.delete(env_dir)
 
     # ----------------------------------
     def create_snapshot_dir(self,snapshot_name):
         env_dir = self.config_files_root + '/Snapshots/' + snapshot_name.strip()
-        if not self.storage_client.dir_exist(env_dir):
-            self.storage_client.create_dir(env_dir, write_to_output=True)
+        if not self.storage_mgr.dir_exist(env_dir):
+            self.storage_mgr.create_dir(env_dir, write_to_output=True)
 '''
